editoast: code refactoring of the redis_utils module

+++ editoast/src/map/layers.rs
@@ -18,7 +18,7 @@ pub struct View {
     pub exclude_fields: Vec<String>,
     #[serde(default)]
     pub joins: Vec<JoinExpr>,
-    pub cache_duration: u32,
+    pub cache_duration: u64,
     #[serde(rename = "where", default)]
     pub where_expr: Vec<String>,
 }
